Fixed calculations for Value_Iterator.cpp   *Utilities for all states were not set to 0 initially   *Terminal states had an uninitialized actions map_test     -Added a terminal state action to correct utility calculations      for said states     -Could be a cleaner way to implement. Looking into it.

+++ ProgAssign_4/MDP.h
@@ -19,7 +19,7 @@
 typedef std::pair<int, float> Next_State;
 
 struct State{
-  int reward;
+  float reward;
   float utility;
   // char = action_id (S, G, B)
   // actions = all possible result states for taking action_id
